// frontend/src/adapter/api/__generated/api.ts
export interface PostShoppingListsRequest {
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequest
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequest
     */
    'description'?: string;
    /**
     * 
     * @type {Array<PostShoppingListsRequestItemsInner>}
     * @memberof PostShoppingListsRequest
     */
    'items'?: Array<PostShoppingListsRequestItemsInner>;
}
/**
 * 
 * @export
 * @interface PostShoppingListsRequestItemsInner
 */
export interface PostShoppingListsRequestItemsInner {
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequestItemsInner
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequestItemsInner
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostShoppingListsRequestItemsInner
     */
    'description'?: string;
}